+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/demandresponseandloadcontrol/CancelAllLoadControlEvents.java
@@ -95,7 +95,7 @@ public class CancelAllLoadControlEvents extends ZclDemandResponseAndLoadControlC
 
     @Override
     public void deserialize(final ZclFieldDeserializer deserializer) {
-        cancelControl = (Integer) deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
+        cancelControl = deserializer.deserialize(ZclDataType.BITMAP_8_BIT);
     }
 
     @Override
