+++ com.zsmartsystems.zigbee/src/main/java/com/zsmartsystems/zigbee/zcl/clusters/groups/GetGroupMembershipResponse.java
@@ -136,11 +136,11 @@ public class GetGroupMembershipResponse extends ZclGroupsCommand {
         // Create lists
         groupList = new ArrayList<Integer>();
 
-        capacity = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
-        Integer groupCount = (Integer) deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        capacity = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
+        Integer groupCount = deserializer.deserialize(ZclDataType.UNSIGNED_8_BIT_INTEGER);
         if (groupCount != null) {
             for (int cnt = 0; cnt < groupCount; cnt++) {
-                groupList.add((Integer) deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER));
+                groupList.add(deserializer.deserialize(ZclDataType.UNSIGNED_16_BIT_INTEGER));
             }
         }
     }
